Revert "REST, TW-71471: Performance fix, review fix: Do not invent the wheel (StringPool), use an existing one instead of a writing new code."
Getting back to a separate StringPool implementation as StringPoolInstance tries to intern every unique incoming string. We try to use pooled string for Locator.myRawValue, which means that every single possible locator string will be interned. This is slow and will likely pollute the internal pool which is all kinds of bad.

This reverts commit 0cfa5e97